+++ src/main/java/cn/milai/ibdemo/character/UltraSide.java
@@ -3,9 +3,13 @@ package cn.milai.ibdemo.character;
 import cn.milai.ib.character.AbstractIBCharacter;
 import cn.milai.ib.container.ui.UIContainer;
 
+/**
+ * 乌托拉侧面
+ * @author milai
+ */
 public class UltraSide extends AbstractIBCharacter {
 
-	public UltraSide(int x, int y, UIContainer container) {
+	public UltraSide(double x, double y, UIContainer container) {
 		super(x, y, container);
 	}
